+++ package.json
@@ -39,7 +39,7 @@
     "eslint": "^8",
     "eslint-config-next": "15.1.2",
     "postcss": "^8",
-    "prisma": "^6.0.1",
+    "prisma": "^6.2.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
