+++ src/app/types/exam.ts
@@ -1,0 +1,12 @@
+export interface ExamData {
+    ลำดับ: number;
+    วิชา: string;
+    กลุ่ม: string;
+    'ชั้นปี': string;
+    'นศ.': string;
+    เวลา: string;
+    'ผู้สอน': string;
+    อาคาร: string;
+    ห้อง: string;
+    หมายเหตุ?: string;
+  }
